+++ src/Resources/app/administration/src/main.js
@@ -6,6 +6,7 @@ import './module/heptacom-admin-open-auth-client';
 import './provider/cidaas/heptacom-admin-open-auth-client-edit-page';
 import './provider/google_cloud/heptacom-admin-open-auth-client-edit-page';
 import './provider/jira/heptacom-admin-open-auth-client-edit-page';
+import './provider/jumpcloud/heptacom-admin-open-auth-client-edit-page';
 import './provider/keycloak/heptacom-admin-open-auth-client-edit-page';
 import './provider/microsoft_azure/heptacom-admin-open-auth-client-edit-page';
 import './provider/microsoft_azure_oidc/heptacom-admin-open-auth-client-edit-page';